rename inference headers to align cuda version
