+++ intel_extension_for_deepspeed/op_builder/csrc/transformer/inference/csrc/inference_onednn_wrappers.cpp
@@ -1,4 +1,4 @@
-#include "onednn_wrappers.hpp"
+#include "inference_onednn_wrappers.hpp"
 #include "inference_sycl_layers.h"
 #include <oneapi/dnnl/dnnl_sycl.hpp>
